+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @spoosh/angular
 
+## 0.4.0
+
+### Breaking Changes
+
+- Updated for @spoosh/core@0.7.0 unified tags API
+- Removed `additionalTags` option (use `tags` array instead)
+
 ## 0.3.0
 
 ### Breaking Changes